% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addDiffResults}
\alias{addDiffResults}
\title{Add Differential Analysis Results}
\usage{
addDiffResults(object, results, name)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{results}{Data.frame or list of differential analysis results}

\item{name}{Character. Name for the differential analysis results}
}
\value{
Updated BulkRNAseq object
}
\description{
Add differential analysis results to BulkRNAseq object
}
