% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar}
\alias{bar}
\alias{bar,BulkRNAseq-method}
\title{Bar Plots for Gene Expression}
\usage{
bar(
  object,
  data = "data",
  genes = "CD36",
  group = "group",
  palette = c("#00AFBB", "#E7B800"),
  method = "t.test"
)

\S4method{bar}{BulkRNAseq}(
  object,
  data = "data",
  genes = "CD36",
  group = "group",
  palette = c("#00AFBB", "#E7B800"),
  method = "t.test"
)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{data}{Character, data source for expression data or name of gsva result}

\item{genes}{Character vector, gene names to plot}

\item{group}{Character, column name in metadata for grouping}

\item{palette}{Character vector, colors for groups}

\item{method}{Character, statistical test method for comparison}
}
\value{
A ggplot object
}
\description{
Create bar and box plots for gene expression data from BulkRNAseq objects.
}
\examples{
\dontrun{
data(counts)
data(metadata)
bulk_obj <- BulkRNAseq(counts, metadata)
bar(bulk_obj, genes = "CD36")
bar(bulk_obj, genes = c("CD36", "GAPDH"))
}
}
