% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4class_def.R
\docType{class}
\name{BulkRNAseq-class}
\alias{BulkRNAseq-class}
\title{BulkRNAseq S4 Class}
\description{
An S4 class for storing and analyzing bulk RNA-seq data
}
\section{Slots}{

\describe{
\item{\code{counts}}{matrix. Raw count matrix (genes x samples)}

\item{\code{data}}{matrix. Normalized data matrix (genes x samples)}

\item{\code{metadata}}{data.frame. Sample metadata (rownames are sample names)}

\item{\code{allDiff}}{list. Differential analysis results (list of data.frames)}

\item{\code{geneSet}}{list. Gene sets for GSEA/GSVA analysis (multiple data.frames)}

\item{\code{gsea}}{list. GSEA/enrichment analysis results (data.frames)}

\item{\code{gsva}}{list. GSVA/ssGSEA analysis results (pathways x samples data.frames)}
}}

