% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{colnames,BulkRNAseq-method}
\alias{colnames,BulkRNAseq-method}
\title{Get Column Names}
\usage{
\S4method{colnames}{BulkRNAseq}(x)
}
\arguments{
\item{x}{A BulkRNAseq object}
}
\value{
Character vector of sample names
}
\description{
Get sample names (column names of count matrix)
}
