% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim,BulkRNAseq-method}
\alias{dim,BulkRNAseq-method}
\title{Get Dimensions}
\usage{
\S4method{dim}{BulkRNAseq}(x)
}
\arguments{
\item{x}{A BulkRNAseq object}
}
\value{
Integer vector of length 2 (nrow, ncol)
}
\description{
Get dimensions of the count matrix
}
