% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4class_def.R
\name{.add_msigdb_genesets}
\alias{.add_msigdb_genesets}
\title{Add MSigDB Gene Sets}
\usage{
.add_msigdb_genesets(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
BulkRNAseq object with gene sets added
}
\description{
Internal function to add MSigDB gene sets to BulkRNAseq object
}
\keyword{internal}
