% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCounts}
\alias{getCounts}
\title{Get Counts Matrix}
\usage{
getCounts(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
Matrix of raw counts
}
\description{
Extract the raw counts matrix from BulkRNAseq object
}
