% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getData}
\alias{getData}
\title{Get Normalized Data Matrix}
\usage{
getData(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
Matrix of normalized data
}
\description{
Extract the normalized data matrix from BulkRNAseq object
}
