% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get Gene Sets}
\usage{
getGeneSets(object, name = NULL)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{name}{Character. Name of the gene set to extract.
If NULL, returns all gene sets.}
}
\value{
List or data.frame of gene sets
}
\description{
Extract gene sets from BulkRNAseq object
}
