% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMetadata}
\alias{getMetadata}
\title{Get Metadata}
\usage{
getMetadata(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
Data.frame of sample metadata
}
\description{
Extract the metadata from BulkRNAseq object
}
