% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hmap}
\alias{hmap}
\alias{hmap,BulkRNAseq-method}
\title{Heatmap Visualization}
\usage{
hmap(
  object,
  data = "data",
  genes = "CD36",
  group = "group",
  show_rownames = TRUE,
  scale = "row",
  color_palette = NULL
)

\S4method{hmap}{BulkRNAseq}(
  object,
  data = "data",
  genes = "CD36",
  group = "group",
  show_rownames = TRUE,
  scale = "row",
  color_palette = NULL
)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{data}{Character, data source: "data" for expression data or name of gsva result}

\item{genes}{Character vector of gene names or numeric for top differential genes}

\item{group}{Character, column name in metadata for annotation}

\item{show_rownames}{Logical, whether to show gene names}

\item{scale}{Character, scaling method ("row", "column", "none")}

\item{color_palette}{Color palette for heatmap}
}
\value{
A pheatmap object
}
\description{
Create heatmaps for gene expression data from BulkRNAseq objects.
}
\examples{
\dontrun{
hmap(bulk_obj, genes = c("CD36", "GAPDH"))
hmap(bulk_obj, genes = 10)  # Top 10 differential genes
}
}
