% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MA}
\alias{MA}
\alias{MA,BulkRNAseq-method}
\title{MA Plot}
\usage{
MA(
  object,
  data = "allDiff",
  fc_threshold = 1,
  adj.P_value_threshold = 0.05,
  highlight_fc_threshold = 4,
  point_size = 1.2,
  highlight_size = 3,
  colors = c("darkred", "grey50", "royalblue4", "green4")
)

\S4method{MA}{BulkRNAseq}(
  object,
  data = "allDiff",
  fc_threshold = 1,
  adj.P_value_threshold = 0.05,
  highlight_fc_threshold = 4,
  point_size = 1.2,
  highlight_size = 3,
  colors = c("darkred", "grey50", "royalblue4", "green4")
)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{data}{Character, name of differential analysis result}

\item{fc_threshold}{Numeric, fold change threshold}

\item{adj.P_value_threshold}{Numeric, adjusted p-value threshold}

\item{highlight_fc_threshold}{Numeric, fold change threshold for highlighting}

\item{point_size}{Numeric, size of points}

\item{highlight_size}{Numeric, size of highlighted points}

\item{colors}{Character vector, colors for different groups}
}
\value{
A ggplot object
}
\description{
Create MA plots for differential expression results.
}
\examples{
\dontrun{
MA(bulk_obj, data = "allDiff")
}
}
