% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pca_hc}
\alias{pca_hc}
\alias{pca_hc,BulkRNAseq-method}
\title{PCA and Hierarchical Clustering Plot}
\usage{
pca_hc(
  object,
  dist_method = c("euclidean", "manhattan"),
  hclust_method = c("ward.D2", "complete", "average"),
  k = 2,
  horiz = FALSE,
  k_colors = "lancet",
  rect = TRUE,
  rect_fill = TRUE,
  pca_style = "default",
  add_ellipses = TRUE,
  repel = TRUE
)

\S4method{pca_hc}{BulkRNAseq}(
  object,
  dist_method = c("euclidean", "manhattan"),
  hclust_method = c("ward.D2", "complete", "average"),
  k = 2,
  horiz = FALSE,
  k_colors = "lancet",
  rect = TRUE,
  rect_fill = TRUE,
  pca_style = "default",
  add_ellipses = TRUE,
  repel = TRUE
)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{dist_method}{Distance method for clustering}

\item{hclust_method}{Hierarchical clustering method}

\item{k}{Number of clusters}

\item{horiz}{Logical, horizontal dendrogram}

\item{k_colors}{Color palette for clusters}

\item{rect}{Logical, draw rectangles around clusters}

\item{rect_fill}{Logical, fill rectangles}

\item{pca_style}{PCA plot style}

\item{add_ellipses}{Logical, add confidence ellipses}

\item{repel}{Logical, use repel for labels}
}
\value{
A combined plot object
}
\description{
Create combined PCA and hierarchical clustering plots.
}
\examples{
\dontrun{
pca_hc(bulk_obj)
}
}
