% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rownames,BulkRNAseq-method}
\alias{rownames,BulkRNAseq-method}
\title{Get Row Names}
\usage{
\S4method{rownames}{BulkRNAseq}(x)
}
\arguments{
\item{x}{A BulkRNAseq object}
}
\value{
Character vector of gene names
}
\description{
Get gene names (row names of count matrix)
}
