% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setData}
\alias{setData}
\title{Set Normalized Data}
\usage{
setData(object, data)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{data}{Matrix of normalized data}
}
\value{
Updated BulkRNAseq object
}
\description{
Set the normalized data matrix in BulkRNAseq object
}
