% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,BulkRNAseq-method}
\alias{show,BulkRNAseq-method}
\title{Show BulkRNAseq Object}
\usage{
\S4method{show}{BulkRNAseq}(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
Invisibly returns the object (for method chaining)
}
\description{
Display method for BulkRNAseq objects
}
