% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,BulkRNAseq-method}
\alias{[,BulkRNAseq-method}
\title{Subset BulkRNAseq Object}
\usage{
\S4method{[}{BulkRNAseq}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A BulkRNAseq object}

\item{i}{Row indices (genes) to subset}

\item{j}{Column indices (samples) to subset}

\item{drop}{Logical. Whether to drop dimensions (not used, for compatibility)}
}
\value{
A subsetted BulkRNAseq object
}
\description{
Subset BulkRNAseq object by genes and/or samples
}
