% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BulkRNAseq}
\alias{summary.BulkRNAseq}
\title{Summary of BulkRNAseq Object}
\usage{
\method{summary}{BulkRNAseq}(object)
}
\arguments{
\item{object}{A BulkRNAseq object}
}
\value{
Invisibly returns a list with summary information
}
\description{
Provide a detailed summary of BulkRNAseq object contents
}
