% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{volcano}
\alias{volcano}
\alias{volcano,BulkRNAseq-method}
\title{Volcano Plot}
\usage{
volcano(
  object,
  data = "allDiff",
  fc_threshold = 1,
  highlight_fc_threshold = 3,
  adj.P_value_threshold = 0.05,
  point_size = 1.2,
  highlight_size = 3,
  colors = c("darkred", "black", "royalblue4", "green2"),
  symmetry = FALSE
)

\S4method{volcano}{BulkRNAseq}(
  object,
  data = "allDiff",
  fc_threshold = 1,
  highlight_fc_threshold = 3,
  adj.P_value_threshold = 0.05,
  point_size = 1.2,
  highlight_size = 3,
  colors = c("darkred", "black", "royalblue4", "green2"),
  symmetry = FALSE
)
}
\arguments{
\item{object}{A BulkRNAseq object}

\item{data}{Character, name of differential analysis result}

\item{fc_threshold}{Numeric, fold change threshold}

\item{highlight_fc_threshold}{Numeric, fold change threshold for highlighting genes}

\item{adj.P_value_threshold}{Numeric, adjusted p-value threshold}

\item{point_size}{Numeric, size of points}

\item{highlight_size}{Numeric, size of highlighted points}

\item{colors}{Character vector, colors for up, non-significant, down, and highlighted genes}

\item{symmetry}{Logical, whether to make x-axis symmetric}
}
\value{
A ggplot object
}
\description{
Create volcano plots for differential expression results.
}
\examples{
\dontrun{
volcano(bulk_obj, data = "allDiff")
}
}
